Don't try to copy printers.conf if we haven't created one...
git-svn-id: svn+ssh://src.apple.com/svn/cups/cups.org/trunk@8182 7a7537e8-13f0-0310-91df-b6672ffda945

+++ test/run-stp-tests.sh
@@ -384,7 +384,11 @@ EOF
 	i=`expr $i + 1`
 done
 
-cp /tmp/cups-$user/printers.conf /tmp/cups-$user/printers.conf.orig
+if test -f /tmp/cups-$user/printers.conf; then
+	cp /tmp/cups-$user/printers.conf /tmp/cups-$user/printers.conf.orig
+else
+	touch /tmp/cups-$user/printers.conf.orig
+fi
 
 #
 # Setup the paths...
